Cast REMEMBER_ME_DAYS env value to integer
The env() helper always returns strings, causing RememberMeTest to fail
when comparing against integer 30. Cast the value to int in the config.

Fixes: Failed asserting that '30' is identical to 30

+++ config/auth.php
@@ -78,7 +78,7 @@ return [
     */
 
     'remember' => [
-        'duration' => env('REMEMBER_ME_DAYS', 30),
+        'duration' => (int) env('REMEMBER_ME_DAYS', 30),
     ],
 
 ];
